qt - Split AsyncRenderer code into a separate file
