+++ src-qt/main.cpp
@@ -2,12 +2,13 @@
 #include "cfdg_highlighter.h"
 #include <QApplication>
 #include <QTextEdit>
+#include <QStringList>
 #include <QtDebug>
 
 int main(int argc, char *argv[])
 {
     if (QIcon::themeName() == "") {
-        QIcon::setThemeSearchPaths((QStringList){"/usr/share/icons"});
+        QIcon::setThemeSearchPaths((QStringList) {"/usr/share/icons"});
         QIcon::setThemeName("Tango");
     }
     QApplication a(argc, argv);
